aalok/requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argbind==0.3.9
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
av==14.3.0
boto3==1.38.10
botocore==1.38.10
cached-property==2.0.1
cached_path==1.7.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chex==0.1.89
click==8.1.8
coloredlogs==15.0.1
contourpy==1.2.1
ctranslate2==4.6.0
cycler==0.12.1
datasets==3.5.1
debugpy==1.8.1
decorator==5.2.1
descript-audio-codec==1.0.0
dill==0.3.8
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.1
einx==0.3.0
elevenlabs==1.57.0
ema-pytorch==0.7.7
encodec==0.1.1
etils==1.12.2
executing==2.2.0
fastapi==0.115.12
faster-whisper==1.1.1
ffmpy==0.5.0
filelock==3.13.1
fire==0.7.0
flatbuffers==25.2.10
flatten-dict==0.4.2
flax==0.10.6
fonttools==4.53.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.0rc0
google-auth==2.40.1
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.13.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
humanize==4.12.3
hydra-core==1.3.2
idna==3.10
importlib_resources==6.5.2
inquirerpy==0.3.4
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jax==0.6.0
jaxlib==0.6.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.0
jsonpointer==3.0.0
julius==0.2.7
jupyter-lsp==2.2.5
jupyterthemes==0.20.0
kiwisolver==1.4.5
lazy_loader==0.4
lesscpy==0.15.1
librosa==0.11.0
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.8
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
noisereduce==3.0.3
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.21.1
openai-whisper==20240930
openwakeword==0.4.0
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.13
packaging==25.0
pandas==2.2.3
parso==0.8.4
pendulum==3.1.0
pexpect==4.9.0
pfzy==0.3.4
pillow==10.3.0
platformdirs==4.3.7
ply==3.11
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyloudnorm==0.1.1
pyparsing==3.1.2
pypinyin==0.54.0
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
randomname==0.2.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.6
setuptools==70.1.0
silero-vad==5.1.2
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorstore==0.1.74
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
toolz==1.0.0
torch==2.7.0+cpu
torch-stoi==0.2.3
torchaudio==2.7.0+cpu
torchdiffeq==0.2.5
tqdm==4.67.1
traitlets==5.14.3
transformers-stream-generator==0.0.5
treescope==0.1.9
triton==3.3.0
twilio==9.6.0
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
vocos==0.1.0
wandb==0.19.10
wcwidth==0.2.13
webcolors==24.6.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
x-transformers==2.3.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
